makefile options to bundle expak and pkg_resources modules
These are the only two external module dependencies, and they should both be
pretty stable. For maximum portability it's probably best to bundle them, at
least in the default/easy choice for distribution.

Maybe it would be good to have a config option to use the external (system
install) version of expak? Or maybe try to import system versions of these
modules first and only fallback to the bundled ones when necessary?

Regardless it would be good to verbose-print about whether system or bundled
versions of these modules are being used, and also print expak version.