switch to a different format for expressing normalized DB
This is more compatible with older versions of SoX.

"yum install sox" on a Centos 6.4 system gave me sox version 14.2, which
doesn't understand "gain -n".

"norm" still works with the newer version 14.4.1, and _should_ be the same as
"gain -n" according to the documentation. But on the newer version of sox, the
files produced using "norm" are not identical as with "gain -n" so I'm not
sure what's up with that.

It also looks like with the newer version it might be better to use the
"--norm" global option rather than these.

I'm also seeing more clipping using the older version of SoX (on Linux). Not
sure how to best tackle all this in the documentation etc. ... can I just tell
Linux users to figure out how to get the newer version?

At the very least, the docs should mention that the default command args
might not work with older versions of the utilities.